Added md5sum column for Message IDs in Article table.
MessageIDs can be very long strings, at least in theory. This is why
the corresponding field in the article table is defined as to be
quite long.

The problem is, an index on a large field is not usable in MySQL.
We got feedback from a MySQL optimization consultant we should
remove this.

Also, on SQL Server databases, the index on the message_id column
can potentially cause trouble, because the column is too big.
This would mean some (larger) message IDs can not be inserted.

This patch fixes the issue by introducing a column that stores
the md5sum of the message id, and by introducing an index on
this much smaller column.

Apart from that, when implementing this, I found the
message_id was also present in the article_search_index table
but was not used at all, so I dropped this column and
corresponding index from the table.