Added -a reinstall-all feature to bin/otrs.PackageManager.pl
If you upgraded your OTRS system, it can be that the RPM or Windows
Installer replaced files that were modified by any packages you had
installed. With this new option you can reinstall all packages
with one command - this also allows to integrate it in the
upgrade RPM procedure itself.