Made 'auto-reply' optional for Phone/Custom tickets
As mentioned in IdeaScale # 366 Option to suppress auto responses to new call tickets
https://otrsteam.ideascale.com/a/dtd/Option-to-suppress-auto-responses-to-new-call-tickets/104579-10369

SystemConfig Core::PostMaster has now the setting:

AutoResponseForWebTickets

Defaults: YES # to be backward compatible

If this option is set to 'Yes', tickets created via the web interface, via Customers or Agents, will receive an autoresponse if configured. If this option is set to 'No', no autoresponses will be sent.