Added dropdown of available transition actions in dialog.
Today I'm attending Process Management training.
I found it very unpleasant to have to type in

  Kernel::System::ProcessManagement::TransitionAction::TicketStateSet

every time I want to add a transition action. It's a long string and it's error prone.
I added a selection that shows a list of available transition actions.

Of course you still need to know what key-value pairs are needed as arguments for the
transition action, but at least entering a new transition action now is a little more
convenient.