Fixed Ticket Loop protection with multiple From addresses.
Fix loop protection error on PostgreSQL and Oracle if auto reponse is set
and more than one customer is selected in Phone ticket.

Since OTRS 3.1 the FROM field in articles can contain more then one
address, but only if used from the Phone ticket mask.
In this case the complete FROM field with all recipients is stored
in the loop protection table; and is quite easy to overflow. On
MySQL this is silently truncated but on PostgreSQL or Oracle it
does not, and causes an error.

This patch splits out the addresses and inserts a single line for
every address, which is more clean and mitigates the size issue.

BTW loop protection is only used for auto-replies at the moment.

Fixes bug#9594 - http://bugs.otrs.org/show_bug.cgi?id=9594
'No auro-reply sent with multiple From addresses in
AgentTicketPhone on PostgreSQL and Oracle.'