Only add language files for languages on the system
Especially on systems with just a few languages installed this will
really decrease the size of the apache process.

Also removed UnitTest files from the child, mainly becuase it depends
on WWW::Selenium which is not present on most customers systems.

And I fixed escaping on some $ signs so the generated startup file
is OK again.