Set Cookie-Path for OTRS cookies.
Setting the path for a cookie restricts its use to the given path. This
provides some extra security, other tools running on the same host but
using a different path can no longer access OTRS's cookies.
Because often OTRS is deployed on one host that does nothing but OTRS,
the extra security is limited, but still I feel we should use this
option.

See:
  - http://en.wikipedia.org/wiki/HTTP_cookie#Domain_and_Path
  - http://tools.ietf.org/html/rfc6265#section-5.2.4