Refactored mail backends.
We now use Net::SSLGlue::POP3 and Net::SSLGlue::SMTP which allow
us to use the same base modules for the SSLified backends as for
the 'plain' backends.

Therefore we drop the following modules:

 - Net::SMTP::SSL
 - Net::SMTP::TLS::ButMaintained
 - Mail::POP3Client
 - Net::POP3::SSLWrapper

Because we now use similar mail libraries for the connections, we can
use much less code; and we use 'use base' to derive the content of
the SMTP, IMAP, and POP3 backends in the SSLified versions.

Another benefit of using the Net::SSLGlue modules is that we can force
SSL certificate verification to 'off' - or we can pass verification
options to IO::Socket::SSL if we want to do this in the future.
Passing full options to IO::Socket::SSL was previously not possible
with all mail modules.