Always display Scheduler warning if it is not active.
The OTRS Scheduler Service should now always be running, and not only
if there is a web service configured. This is because it will be used
by more parts of OTRS than just the Generic Interface. The notification
test to see if the Scheduler is running, is now always run.

Also, if you're not in the Admin group, you'll also see the
notification, but you can't start the service from the UI. Instead,
you're instructed to warn your administrator.