Follow-up fix for bug#9541 - HTTPS via proxy.
Net::HTTP had a bug that caused issues when used with a proxy server.
This is fixed in version 6.03 and up. However, some distributions,
most notably Ubuntu 12.04 LTS, ship 6.02, rendering them not able to
use HTTPS via a proxy.

Ubuntu 10.04 has much older versions of Net::HTTP that are not
affected.

See also this launchpad bug:
https://bugs.launchpad.net/ubuntu/+source/libnet-http-perl/+bug/1057609

We now bundle Net::HTTP 6.06 in Kernel/cpan-lib in order to be sure
we're not affected by this issue.