Rollback changes for bug#9613.
The fix causes problems if you have modified for instance
AgentTicketPhone.pm but not AgentTicketPhone.dtl or the other way
around.
Also if you would have put AgentTicketPhone.dtl in a theme you'd
it'd cause problems.
We keep the fix in release 3.3 because there you'd need to port
the template and perl files anyway.