Fixed message when message is too large.
First of all, the message size is in bytes and the maximum message size
in SysConfig is in kilobytes - I made a division error so the error
message you see would be:

IMAPS: Can't fetch email .. Email too big (17 MB - max 0 MB)!

And then I realized that converting it to megabytes might be good
for readability but BAD for troubleshooting, it could very well
be that the message size is 16783 KB and the maximum is 16384,
the error message would then be:

Email too big (16 MB - max 16 MB)!
which is not very helpful. So now it displays the sizes in KB again.