Added config opt DisableMSIFrameSecurityRestricted
The security=restricted iframe option that we have in OTRS since 3.1
causes problems in certain conditions.
One such scenario is when used with single sign on and IE8 on Windows
XP.
The other scenario is when used with a specific Juniper SSL webvpn
appliance; this appliance uses a third-party cookie for authentication
of the user. security=restricted does not allow the third-party
cookie, logging the user out and showing the SSL appliance login
page instead of the iframe contents.

These are edge cases, but for these cases we now add this
configuration option.