Replaced mail based registration in installer.
This changes the mail based registration with a registration in the
Admin frontend.

It allows you to register your system with the OTRS Portal. On the OTRS
Portal you'll get an overview of the registered systems against your
Customer ID.

After registration, your system will send updates to the OTRS Portal so
the overview of your registered systems with their OTRS versions will
be kept up-to-date.

Registration is of course optional, you can also use OTRS without
registering.