Fixed customersearch with non-ascii chars on MSSQL
This fixes bug#7008 - http://bugs.otrs.org/show_bug.cgi?id=7008

On SQL Server, the CustomerUser.t and Service.t tests were failing.
This is because using a SQL Statement like this:

SELECT id FROM customer_user WHERE login LIKE '%string%';

causes problems if the string contains non-ascii characters on SQL
Server. This is a problem with DBD::ODBC actually.

It DOES work if we use bind-variables in the SQL statement.
Since using bind-variables is better for multiple reasons,
including preventing SQL injection, we changed the code to use
bind variables.

For this we extended the DBObjects' QueryCondition
to include a mode where it returns both the SQL string and the
bind variables.