Allow to set 0 TimeUnits in TicketCreate WS
If Ticket::Frontend::NeedAccountedTime is enabled, a time unit MUST be
specified before being able to create a ticket.

Currently, you can not set the TimeUnit to '0'; this fixes it.
I would consider it even better if the web service would use
a default of '0' if no value is passed.

This is also because the TimeUnit is only required in the Agent
Interface, for Agent articles; and with the TicketCreate web service
you can also create customer (or system) articles.