Fixed: back-action in searches with 1 result.
This fixes bug#8991.
http://bugs.otrs.org/show_bug.cgi?id=8991

If a search result finds just one ticket, it opens the ticket
directly instead of presenting the single result.

This is helpful, but sometimes you want to go back to the search
result to for instance click 'Change Search Options'. The
Back button in the ticket zoom did not work properly in this
scenario. This small patch fixes the problem.