Don't show Apache modules on Windows.
On Windows, we either deploy on IIS + PerlEx (recommended) or on
Apache + CGI and no longer as mod_perl because of stability issues.
CheckModules still showed errors in the Support Module if mod_perl
was not installed, which is NOT needed. So on Windows we now no
longer show Apache:: modules and mod_perl and if these modules
are missing or not because it's not relevant.

Also, the 'use perl -MCPAN -e shell' instructions that were displayed
if a module was not installed is typically not the best way to install
a perl module. The best way is to get them from your package repository
and IF you need to use cpan you'd just type 'cpan Some::Module'.

Rolf is working on a nice way to use checkmodules to install the
packages from your package manager.