Fix: customer data source in non-utf8 does not work.
See http://bugs.otrs.org/show_bug.cgi?id=10005

if you have a remote data source in non-utf8 the data was not
presented correctly.

the fix is to make the database library convert the data for you;
actually MySQL already did that for us. This caused double conversion
problems.
No conversion is needed in the CustomerUser::DB and CustomerCompany::DB
backends.