Revert "Using DBD utf8 switch for mysql"
This reverts commit b7e7df17024491526a324832ca286fa60020ba1d.

After this change we just had 1 test failing: Webservice.t

The problem is that if we use mysql_enable_utf8 only columns that
are defined as 'char', 'varchar', and 'text' have the utf-8 flag
on when retrieved from the database.

the config column in the web service history is not defined as
text but as a 'longblob':

    CREATE TABLE gi_webservice_config (
    id INTEGER NOT NULL AUTO_INCREMENT,
    name VARCHAR (200) NOT NULL,
    config LONGBLOB NOT NULL,
    ...

This retults in the config data returning to the database not to
have the utf-8 flag on.

If we'd upgrade the column to LONGTEXT, unittests will pass:

    ALTER TABLE gi_webservice_config CHANGE config config LONGTEXT NOT NULL;

But we actually have quite a lot columns defined as LONGBLOB where
they really should be LONGTEXT so I'll revert this change now.