TicketAcl: Only fetch required data
This analyzes ACLs for the information they need to match,
and only fetches the required pieces of information.

This should speed up real-world ACL usage quite a bit,
Especially when ACL don't control all aspects (state, priority, type,
service, process, CustomerUser, SLA etc.) of the system.

On the other hand it requires that ACL *modules* pre-declare
the aspects or Checks that they need as arguments.