Currently QueueView only allows to navigate to queues which have at least one unlocked ticket. Once the queue becomes visible, it is well possible to switch between filters that show either All open or just Unlocked tickets in that queue.
Encouraged by very positive feedback in the bugzilla ticket http://bugs.otrs.org/show_bug.cgi?id=8244#c4, this PR modifies the behaviour as follows:

All readable queues have both non-closed and unlocked tickets enumerated. The results are presented in the "Queue Bar" in the same order as the filters. This overview continues to be generated through the the IndexAccelerator backends which have been modified accordingly. Their queries now count tickets grouped by lock, and the filtering/consolidation is done in code, so they can return a complementary 'Total' field in addition to the existing 'Count'.

A bug that produced incorrect sum of the number of tickets in nested queues has also been fixed.