Dynamic field updates should pass OldValue to Event and History
Currently the value prior to an update of a dynamic field
can only be recalculated from history which is fairly expensive.

Supplying the "OldValue" as part of the TicketDynamicFieldUpdate_*
event allows an event handler to react on certain _changes_ in value
(e.g. increase, decrese) without digging into the history.
Recovery from history again is eased by adding the OldValue as
an extra item in the "Name" field.