Fix bug 10213: Internet Explorer does not provide JSON.parse
Non-ancient versions of Internet Explorer do in fact provide JSON.parse,
but do not expose that method in compatibility mode.

The rules by which IE decides to enable which compatibility mode are obscure,
so this commit forces "edge", which means it enables all features that
are available. At least on IE 11 this makes the error message go away,
but I suspect it will actually fix IE8+