OTRS uses Mail::Address, not Email::Address.
I found this when I was checking to see if OTRS would be vulnerable
for CVE-2014-0477: "Email::Address: Denial-of-Service in
Email::Address::parse"

ref: http://seclists.org/oss-sec/2014/q2/563

So, the good news is, it's not vulnerable. The bad news is, I found
this comment where it almost LOOKS as if OTRS is using this module.

:D