Correctly handle UserSessionStart on AuthSessionFS
On AuthSessionFS the UserSessionStart value was not used.

This fixes http://bugs.otrs.org/show_bug.cgi?id=10617
"Using AuthSessionFS, sessions will always be expired"

Also, I improved the perldoc a little on the data structures
returned, and added a unit test for this specific bug.

It might be good to still add some more tests for expired
sessions.