fix bug: Ticket owner was not shown regardless what was configured
When AJAX support for CustomerTicketZoom was added, code was copied from existing Agent frontend modules
without any review. The then new hash %ACLCompatGetParam had a key OwnerID which value was always 'undef'
as it tried to get a value from %GetParam that was never set.
The values of %ACLCompatGetParam overrode the values of %Ticket (in fact it was only OwnerID) so that
the owner could not be shown in CustomerTicketZoom.