Test that classes declare all dependencies
This creates all objects in isolation, checking the state of the object
manager before and after each build; this makes the test a bit ugly,
because it pokes in the OM internals.

One thing this test reveals is that currently, dependencies that depend on
configuration options are not declared. I maintain that they should be
declared, because they are not built automatically anyway, only on demand.