Make strings translatable
Please review all the <Item>...</Item>, <Name>...</Name>, etc. tags, because I think there are many more string which needs to be translated, but they're not.
And please, don't use CamelCase strings, because translators can't decide if it is a value of a parameter or a string to translate. Use separete words instead. For example: instead of "CustomerAgentTicket" use Customer Agent Ticket".