Let LWP::UserAgent handle https-proxy commands
For accessing HTTPS resources via a proxy server, HTTP CONNECT is used
only LWP did not support this properly. See
https://en.wikipedia.org/wiki/HTTP_tunnel#HTTP_CONNECT_tunneling
This is why we had the workaround of using Net::SSL in place.

Only, since LWP 6.06 LWP now supports this properly.
See https://en.wikipedia.org/wiki/HTTP_tunnel#HTTP_CONNECT_tunneling
this means the whole Net::SSL workaround, and the cruft around
this, can be removed!

It also means that previously, OTRS did not verify SSL certificates
when a proxy server was configured, because Net::SSL does not support
this. After this change, SSL certificates are also verified if a proxy
server is configured. And when DisableSSLVerification is enabled,
the certificates are not verified, as before.