Fix Oracle connection in WebInstaller
WebInstaller was unable to create Oracle database.
I found the wrong usage of "$Self->{}" instead of "$DB{}" for DBHost and DBSID, which broke connection to Oracle. On the other hand the correct connection string was stored in ConfigDSN.
Second, the "test database connection" used another naming, requiring the short "SID" in config parameter "DBSID", but the resulting ConfigDSN used the Service Name (in my case the SID extended by or domain name). It should be unique for connection test and real connection and I prefer the Service Name, so I changed the DSN for the connection string.

Maybe someone should care about the naming "SID" now, but I'm fine since it at least works now.

I will add another separate patch to store DBPort and DBSID inside DatabaseHost and Database(name) to avoid having them only in the DSN. But if you don't like it, this patch would be sufficient to make WebInstaller work with Oracle Databases!