Enable to disable to force to rewrap a text input in text area
In recent years, we don't have to rewrap the text in many cases by
following reasons.  This PR enables to disable to force to rewrap a text
in text area as article body.

  - Recent screen is large.  It can display more than 150 characters of
    a line without scroll.
  - Even when send email, body text is encoded quoted-printable or
    base64.  It means that we don't have to mind "Line Length Limits"
    written in RFC 5322 Section 2.1.1.
  - Japanese and Chinese does not use spaces to separate between words
    in locale.  It causes to ignore hard wrap trick in these locales.
    By the way, I confirmed Firefox and Internet Explorer expectly wrap
    a text even in these locales with "word-wrap: break-word".
  - If Frontend::RichText is not "0", a text is not rewrapped.

By the way, I see that the feature to rewrap a text in text area is
introduced in fix for bug#526.  At first "hard" attributes of textarea
is used, it is changed into Perl code in the fix.  However, now, if
"word-wrap: break-word" works correctly, a horizontal scrollbar will be
not caused for a long line, and the line will be wrapped expectedly by
browser's feature.