Update rhel7-otrs-systemd.spec
Moved useradd from "pre" to "post" to avoid conflicting with SELinux.

When Enforcing SELinux, the RPM which created with this spec file makes fail to install. Because creating home directory "/opt/otrs" (equals to install directory) does not allowed in default SELinux policy.
The "/opt/otrs" directory has created in installing process, so it is reasonable than making it to allow.