Added new console interface for a massively improved user and developer experience on the commandline.
otrs.Console.pl is now the central place for all commands (formerly scripts) that use the OTRS API.
You can run "otrs.Console.pl" to get a general overview and a listing of the available commands, or "otrs.Console.pl Command::Name --help" to get detailed help for the individual commands.
Cron scripts are not ported yet because they will be obsolete soon.
Kudos to http://symfony.com/doc/current/components/console/introduction.html for the inspiration and to Marc Bonsels and Carlos Rodriguez for their help with the command porting.