AgentTicketBulk: Redirect to Zoom if only one Ticket comes out
if the bulk action involves only one ticket, or one of the action is to
merge all tickets into one, redirect to the AgentTicketZoom view of
that one ticket after successfully doing the bulk action.

This makes for a much smoother user experience, since usually people
want to actually work with the ticket that they just merged.