Handle /var/lock/subsys/otrs in redhat init script
RedHat <= 6 init scripts need to create a lockfile in /var/lock/subsys or the
init script won't be run on shutdown.

Another requirement is, that one should get more appropriate return codes when
calling the status command. For example puppet only starts the service if it's
not already started (very sane) but right now status returns 0 ("program is
running or service is OK") if start was called or not and in turn does not
enable the cron jobs.