Fixed bug#4661 - SearchProfiles updated on UserLogin
This is a bug fix for bug#4661.
ref: http://bugs.otrs.org/show_bug.cgi?id=4661

Previously, when the login name of an agent or customer user changed,
their search profiles were lost.

I introduced a function in Kernel::System::SearchProfile to update
the search profiles for a given login, plus I modified UserUpdate
and wrote an event module for CustomerUsers.
I also wrote unit tests to cover these scenarios.