In AgentTicketBulk, merge first, and do the other actions only on the main ticket
Some of our users complained that using a bulk action to merge several tickets
into one and setting the owner (for example) at the same time sends a bunch of
redundant notifications.

Merging a ticket negates some changes (for example setting ticket state),
and make others (such as owner changes) irrelevant. So do it first, and
save the other actions for the ticket into which was merged.

If this isn't desirable in the general case, I could make this new behavior
configurable.