Optimize GenericAgent management
The old way was to first query the last execution times, and then deletes them individually.
When several GenericAgent jobs run in parallel (for example in a cluster, and/or when several
E-Mails are delivered into the PostMaster in parallel), this produces a quadratic number
of DELETE statements (which caused an outages for us).

Instead, simply delete all old run time values in a single delete statement, which
keeps the quadratic run time out of OTRS (and confined to the database, which
handles it much better). As a bonus, the code is also shorter and simpler :-).