Bug 8298 I Created a pull request (nielsdimmers:bug8298) with the following changes:
In general, settings are saved, so on almost every admin screen, the “Sumbit” button has been changed to “Save”, with a few exceptions:
- otrs/index.pl?Action=AdminSMIME;Subaction=ShowAddCertificate - I changed the “Submit” button to “Add”, because the user is adding a certificate here
- otrs/index.pl?Action=AdminSysConfig;Subaction=Import - I changed the “Submit” button to “Import”, since the user is importing settings here

Mostly, agent actions are “submitted”, so the submit button can stay whenever there’s a creation of a ticket, adding of an article or sending of a message. Everywhere where there are settings involved (like the settings on the dashboard or in the ticket overview screen) I changed “submit” to “Save” or to “Save Changes” if the last one was more in line with similar situations. There are two exceptions:
- otrs/index.pl?Action=AgentTicketBulk - I changed the “Submit” button to “Excecute Bulk Action”, because the user wants to execute the bulk action
- otrs/index.pl?Action=AgentTicketMove - I changed the “Submit” button to “Move”, since the ticket is moved to a different queue

I propose to close this bug report and whenever there are other changes required, open a new one.