Make unit test output much shorter.
By default STDERR and STDOUT output will be captured and only be shown together with name and number of the failing test case if the test acutally failed. Otherwise only a green dot will appear for every test case.

The old verbose behaviour can be triggered with the flag --verbose.

@Notify: manuel.hecht@otrs.com, gsd@otrs.com