Used File::Spec to construct paths in restore.pl
Use the catfile subroutine from File::Spec to build the paths, instead
of doing "$dir/File.ext". This way, the built paths won't contain double
slashes if the given $dir ends with a slash.

Multiple slashes are a valid separator in Unix-based systems, as far as
I know, but at the very least this is cleaner and the output messages
are less confusing.