Authentication Module Selector Patch
This patch adds ability for administrators to selectively choose authentication modules based on FQDN specified in URL. This works mostly the same as host based skins. The main purpose of this modification is to allow mixed authentication (Form and HTTP Basic/Kerberos) on a single web server deployment, which is not possible in the original OTRS Framework.