Refix Wrong variable <OTRS_CUSTOMER_REALNAME> in AutoResponses #4640
This is a partial revert of 82d55d6 which introcuded a condition where
the From field in an Auto Response would be badly generated. This is documented
in Bug #4640 comments 15 and 16. The field that should be used instead is the To
field and not the From since the generated email already has those headers set