Fix search for dynamic field that contains an asterisk (*)
The TicketSearch code uncondtionally replaced every * with a %, but then
used the "=" operator for comparison, where % is not a wildcard. So
it was impossible to search for a dynamic field value which contains a *.

This commit does the substitution of * by % only if the operator is
"Like", where % is actually a wildcard.