integrate DynamicFieldInteger
Sometimes a textfield is not enough: when you want to save something like storypoints, a rank or something like that
it is much easier to have an integer field. This has also the advantage, that the sorting works as expected. The
text field is sorted asciibetically (e.g. 1, 10, 110, 2, 3, 30, 4, and so on).