Increase file cache speed ~30-40% for small data
FileRead, FileWrite, FileDelete methods work slower if you specify
path to a file with Directory and Filename, because they call
FilenameCleanUp method, which is pretty expensive for cache.
But in FileStorable we don't need FilenameCleanUp, because
cache file names can't have whitespaces, dots, extensions and so on,
as they are just a MD5 digests.